refactor(pyrogue): centralize slice-expression parsing in _HelperFunctions
Move _astNodeToInt() and _parseSliceExpr() from _Variable.py to
_HelperFunctions.py so the safe AST-based slice parser is reusable from
any pyrogue submodule. Replace the eval(f'tmpList[{keys[0]}]') call in
_Node._iterateDict() with _parseSliceExpr(), which also fixes the
negative-index case (e.g. myArray[-1]) that previously fell through to
the broad except branch and silently returned an empty list.
